requirements.txt

azure-functions
pyodbc
pandas
numpy
scikit-learn
